feat(kprobe): Add basic kprobe support for x86_64
This commit extends the TrapFrme method to facilitate use by callback
functions in kprobes. In order to find the function address based on
the symbol, we added the`addr_from_symbol` function.

In the x86 and riscv64 architecture related codes, we have added the
handling of breakpoint exceptions and single-step exceptions.